requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
langchain==0.0.350
langchain-community==0.0.3
langchain-openai==0.0.2
chromadb==0.4.17
wikipedia==1.4.0
openai>=1.6.1,<2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3
sentence-transformers==2.2.2
python-multipart==0.0.6
requests==2.31.0